pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Padlan Alqinsi",email = "padlanalqinsi@live.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "dspy (>=2.6.11,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "ruff (>=0.9.10,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry]
packages = [{ include = "src" }]
